Source data and final report
All data, analysis and final markdown documents (html format). NB code
hidden is easier to read, while code visible has all of the R scripts.
Black text is the problem and questions, red and blue text are the
answers and rationales respectively, and grey boxes are the output
tables and R code.